pyproject.toml
[tool.poetry]
name = "paperreviewer"
version = "0.1.0"
description = ""
authors = ["Navya <navyac.social@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.42.2"
google-generativeai = "^0.8.4"
pypdf = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
